@@ -24,10 +24,10 @@
         "spatie/once": "^2.0"
     },
     "require-dev": {
-        "phpunit/phpunit": "^9.6",
+        "phpunit/phpunit": "^10.0",
         "mockery/mockery": "^1.5",
-        "laravel/framework": "^9.0",
-        "laravel/laravel": "^9.0"
+        "laravel/framework": "^10.0",
+        "laravel/laravel": "^10.0"
     },
     "autoload": {
         "psr-4": {
